requirements.txt
-e git+https://github.com/celtiberi/ai-prd-generator.git@572fc5a652956f01812f72f3413fe5067bc9ccfc#egg=ai_prd_generator&subdirectory=backend
aiohttp==3.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
argcomplete==3.5.3
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.10
distro==1.9.0
ecdsa==0.19.0
faiss-cpu==1.7.4
fastapi==0.109.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
multidict==6.1.0
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.58.1
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pyasn1==0.6.1
pydantic==2.10.4
pydantic-settings==2.1.0
pydantic_core==2.27.2
Pypubsub==4.0.3
pytest==8.0.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.35.1
tavily-python==0.5.0
tiktoken==0.8.0
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.27.0
xmltodict==0.14.2
yarl==1.18.3
yq==3.4.3
